Tadween/langchain_requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
openai==1.12.0
faiss-cpu==1.7.4
flask==3.0.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0 